// generated/src/app/api-models/cap-elig-work.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapEligWork {
  seqCcalcId: number;
  seqCapEligWork: number;
  seqEligHist: number;
  seqMembId: number;
  currentRetro: string;
  addSubtract: string;
  effectiveDate: Date;
  termDate: Date;
  seqGroupId: number;
  planCode: string;
  lineOfBusiness: string;
  medicareStatusFlag: string;
  otherStatusFlag: string;
  seqProvId: number;
  ipaId: string;
  panelId: string;
  seqProv2Id: number;
  userDefined1: string;
  age: number;
  gender: string;
  arcEffectiveDate: Date;
  arcTermDate: Date;
  curEffectiveDate: Date;
  curTerminationDate: Date;
}